refactor: aggregate type definitions into types
