+++ src/admin/pages/Profile/Context/types.ts
@@ -1,6 +1,6 @@
 import { SWRResponse } from 'swr';
 import { SWRMutationResponse } from 'swr/mutation';
-import { Me } from '../../../../configs/types.js';
+import { Me } from '../../../../types/index.js';
 
 export type ProfileContext = {
   updateMe: () => SWRMutationResponse<void, any, string, Record<string, any>>;
